refactor(protocol): extract shared MastForestScript type
